--- app/src/main/java/com/tinkoff/edu/app/LoanCalcRepository.java
@@ -1,5 +1,0 @@
-package com.tinkoff.edu.app;
-
-public interface LoanCalcRepository {
-    int save(LoanRequest loanRequest);
-}
