--- app/src/main/java/com/tinkoff/edu/app/LoanCalcService.java
@@ -1,5 +1,0 @@
-package com.tinkoff.edu.app;
-
-public interface LoanCalcService {
-    LoanResponce createRequest(LoanRequest loanRequest);
-}
