+++ app/src/main/java/com/tinkoff/edu/app/controllers/LoanCalcController.java
@@ -1,4 +1,10 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.controllers;
+
+import com.tinkoff.edu.app.*;
+import com.tinkoff.edu.app.loggers.LoancalcLogger;
+import com.tinkoff.edu.app.repositories.LoanCalcRepository;
+import com.tinkoff.edu.app.services.IpNotFriendlyService;
+import com.tinkoff.edu.app.services.LoanCalcService;
 
 /**
  *
