+++ app/src/main/java/com/tinkoff/edu/app/enums/LoanResponceType.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.enums;
 
 public enum LoanResponceType {
     APPROVE, DECLINE
