+++ app/src/main/java/com/tinkoff/edu/app/enums/LoanType.java
@@ -1,4 +1,4 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.enums;
 
 public enum LoanType {
     PERSON, IP, OOO
