+++ app/src/main/java/com/tinkoff/edu/app/loggers/LoancalcLogger.java
@@ -1,4 +1,6 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.loggers;
+
+import com.tinkoff.edu.app.LoanRequest;
 
 public class LoancalcLogger {
     public static void log(LoanRequest loanRequest) {
