+++ app/src/main/java/com/tinkoff/edu/app/repositories/LoanCalcRepository.java
@@ -1,0 +1,9 @@
+package com.tinkoff.edu.app.repositories;
+
+import com.tinkoff.edu.app.LoanRequest;
+import com.tinkoff.edu.app.LoanResponce;
+import com.tinkoff.edu.app.enums.LoanResponceType;
+
+public interface LoanCalcRepository {
+    LoanResponce save(LoanRequest loanRequest, LoanResponceType responceType);
+}
