+++ app/src/main/java/com/tinkoff/edu/app/services/IpNotFriendlyService.java
@@ -1,4 +1,10 @@
-package com.tinkoff.edu.app;
+package com.tinkoff.edu.app.services;
+
+import com.tinkoff.edu.app.LoanRequest;
+import com.tinkoff.edu.app.LoanResponce;
+import com.tinkoff.edu.app.enums.LoanResponceType;
+import com.tinkoff.edu.app.enums.LoanType;
+import com.tinkoff.edu.app.repositories.LoanCalcRepository;
 
 public class IpNotFriendlyService extends PersonCalcService {
     public IpNotFriendlyService(LoanCalcRepository repo) {
