+++ app/src/main/java/com/tinkoff/edu/app/services/LoanCalcService.java
@@ -1,0 +1,8 @@
+package com.tinkoff.edu.app.services;
+
+import com.tinkoff.edu.app.LoanRequest;
+import com.tinkoff.edu.app.LoanResponce;
+
+public interface LoanCalcService {
+    LoanResponce createRequest(LoanRequest loanRequest);
+}
